removed some files and added restructuring
